Refactor the item sheets for less repetition
This adds a common ForbiddenLandsItemSheet that all other item sheets
uses.
