+++ script/sheet/armor.js
@@ -1,33 +1,11 @@
-export class ForbiddenLandsArmorSheet extends ItemSheet {
+import { ForbiddenLandsItemSheet } from "./item.js";
+
+export class ForbiddenLandsArmorSheet extends ForbiddenLandsItemSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      classes: ["forbidden-lands", "sheet", "item"],
       template: "systems/forbidden-lands/model/armor.html",
       width: 400,
       height: 400,
-      resizable: false,
     });
-  }
-
-  _getHeaderButtons() {
-    let buttons = super._getHeaderButtons();
-    buttons = [
-      {
-        label: "Post Item",
-        class: "item-post",
-        icon: "fas fa-comment",
-        onclick: (ev) => this.item.sendToChat(),
-      }
-    ].concat(buttons);
-    return buttons;
-  }
-
-  getData() {
-    const data = super.getData();
-    return data;
-  }
-
-  activateListeners(html) {
-    super.activateListeners(html);
   }
 }
