+++ script/sheet/building.js
@@ -1,34 +1,12 @@
-export class ForbiddenLandsBuildingSheet extends ItemSheet {
+import { ForbiddenLandsItemSheet } from "./item.js";
+
+export class ForbiddenLandsBuildingSheet extends ForbiddenLandsItemSheet {
 
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
-            classes: ["forbidden-lands", "sheet", "item"],
             template: "systems/forbidden-lands/model/building.html",
             width: 400,
             height: 522,
-            resizable: false
         });
-    }
-
-    getData() {
-        const data = super.getData();
-        return data;
-    }
-
-    _getHeaderButtons() {
-        let buttons = super._getHeaderButtons();
-        buttons = [
-            {
-                label: "Post Item",
-                class: "item-post",
-                icon: "fas fa-comment",
-                onclick: (ev) => this.item.sendToChat(),
-            }
-        ].concat(buttons);
-        return buttons;
-    }
-
-    activateListeners(html) {
-        super.activateListeners(html);
     }
 }
