+++ script/sheet/critical-injury.js
@@ -1,33 +1,11 @@
-export class ForbiddenLandsCriticalInjurySheet extends ItemSheet {
+import { ForbiddenLandsItemSheet } from "./item.js";
+
+export class ForbiddenLandsCriticalInjurySheet extends ForbiddenLandsItemSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      classes: ["forbidden-lands", "sheet", "item"],
       template: "systems/forbidden-lands/model/critical-injury.html",
       width: 400,
       height: 250,
-      resizable: false,
     });
-  }
-
-  getData() {
-    const data = super.getData();
-    return data;
-  }
-
-  _getHeaderButtons() {
-    let buttons = super._getHeaderButtons();
-    buttons = [
-      {
-        label: "Post Item",
-        class: "item-post",
-        icon: "fas fa-comment",
-        onclick: (ev) => this.item.sendToChat(),
-      }
-    ].concat(buttons);
-    return buttons;
-  }
-
-  activateListeners(html) {
-    super.activateListeners(html);
   }
 }
