+++ script/sheet/hireling.js
@@ -1,33 +1,11 @@
-export class ForbiddenLandsHirelingSheet extends ItemSheet {
+import { ForbiddenLandsItemSheet } from "./item.js";
+
+export class ForbiddenLandsHirelingSheet extends ForbiddenLandsItemSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      classes: ["forbidden-lands", "sheet", "item"],
       template: "systems/forbidden-lands/model/hireling.html",
       width: 400,
       height: 400,
-      resizable: false,
     });
-  }
-
-  getData() {
-    const data = super.getData();
-    return data;
-  }
-
-  _getHeaderButtons() {
-    let buttons = super._getHeaderButtons();
-    buttons = [
-      {
-        label: "Post Item",
-        class: "item-post",
-        icon: "fas fa-comment",
-        onclick: (ev) => this.item.sendToChat(),
-      }
-    ].concat(buttons);
-    return buttons;
-  }
-
-  activateListeners(html) {
-    super.activateListeners(html);
   }
 }
