+++ script/sheet/monster-talent.js
@@ -1,34 +1,12 @@
-export class ForbiddenLandsMonsterTalentSheet extends ItemSheet {
+import { ForbiddenLandsItemSheet } from "./item.js";
+
+export class ForbiddenLandsMonsterTalentSheet extends ForbiddenLandsItemSheet {
 
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
-            classes: ["forbidden-lands", "sheet", "item"],
             template: "systems/forbidden-lands/model/monster-talent.html",
             width: 400,
             height: 420,
-            resizable: false
         });
-    }
-
-    getData() {
-        const data = super.getData();
-        return data;
-    }
-
-    _getHeaderButtons() {
-        let buttons = super._getHeaderButtons();
-        buttons = [
-            {
-                label: "Post Item",
-                class: "item-post",
-                icon: "fas fa-comment",
-                onclick: (ev) => this.item.sendToChat(),
-            }
-        ].concat(buttons);
-        return buttons;
-    }
-
-    activateListeners(html) {
-        super.activateListeners(html);
     }
 }
